Fixed double roads and improved traffic lights, modularized the structure.

+++ src/traffic_light.cpp
@@ -1,6 +1,5 @@
 #include "traffic_light.hpp"
 #include <SFML/Graphics.hpp>
-#include <iostream>
 
 TrafficLight::TrafficLight(std::weak_ptr<Road> road, sf::Time green_duration, sf::Time init_time, size_t road_count, State initial_state)
     : m_road(road), m_green_duration(green_duration), m_state(initial_state), m_timer(init_time), m_model(sf::CircleShape(8.f)), adjacent_road_count(road_count)
@@ -29,7 +28,7 @@ void TrafficLight::update(sf::Time elapsed)
         }
         break;
     case State::Red:
-        if (m_timer >= sf::seconds((1.1 * adjacent_road_count) * m_green_duration.asSeconds()))
+        if (m_timer >= sf::seconds((1.1f * static_cast<float>(adjacent_road_count)) * m_green_duration.asSeconds()))
         {
             m_state = State::Green;
             m_timer = sf::Time::Zero;
